backend/requirements.txt
asgiref==3.8.1
awscli==1.33.26
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
docutils==0.16
EB==0.1.5
idna==3.7
jmespath==1.0.1
pathspec==0.10.1
pyasn1==0.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.2
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.1
termcolor==1.1.0
urllib3==1.26.19
wcwidth==0.1.9
